Refactor: 2차 프로젝트 피드백 반영
2차 피드백 일부 반영
