+++ trip_itinerary/src/main/java/com/example/trip_itinerary/itinerary/controller/ItineraryController.java
@@ -1,6 +1,7 @@
 package com.example.trip_itinerary.itinerary.controller;
 
 import com.example.trip_itinerary.itinerary.dto.request.save.ItinerarySaveRequest;
+import com.example.trip_itinerary.itinerary.dto.request.save.StaySaveRequest;
 import com.example.trip_itinerary.itinerary.dto.request.update.ItineraryPatchRequest;
 import com.example.trip_itinerary.itinerary.service.ItineraryService;
 import org.springframework.validation.annotation.Validated;
@@ -16,13 +17,13 @@ public class ItineraryController {
         this.itineraryService = itineraryService;
     }
 
-    @PostMapping("/{id}/itinerary")
-    public Long saveItinerary(@PathVariable Long id, @RequestBody @Validated ItinerarySaveRequest staySaveRequest) {
+    @PostMapping("/{trip_id}/itinerary")
+    public Long saveItinerary(@PathVariable(name = "trip_id") Long id, @RequestBody @Validated ItinerarySaveRequest staySaveRequest) {
         return itineraryService.saveItinerary(id, staySaveRequest);
     }
 
-    @PatchMapping("/itineraries/{id}")
-    public Long patchItinerary(@PathVariable Long id, @RequestBody @Validated ItineraryPatchRequest itineraryPatchRequest) {
+    @PatchMapping("/itineraries/{itinerary_id}")
+    public Long patchItinerary(@PathVariable(name = "itinerary_id") Long id, @RequestBody @Validated ItineraryPatchRequest itineraryPatchRequest) {
         return itineraryService.patchItinerary(id, itineraryPatchRequest);
     }
